+++ react/Reviews.tsx
@@ -1,23 +1,18 @@
 import React, {
-  FunctionComponent,
   Fragment,
-  useContext,
   useEffect,
   useReducer,
 } from 'react'
 import { Helmet } from 'react-helmet'
-import ApolloClient, { ApolloQueryResult } from 'apollo-client'
-import { NormalizedCacheObject } from 'apollo-cache-inmemory'
-import { withApollo } from 'react-apollo'
+import { ApolloQueryResult } from 'apollo-client'
+import { useApolloClient } from 'react-apollo'
 import {
   FormattedMessage,
-  InjectedIntlProps,
-  injectIntl,
   defineMessages,
+  useIntl,
 } from 'react-intl'
-import flowRight from 'lodash.flowright'
-import path from 'ramda/es/path'
-import { ProductContext } from 'vtex.product-context'
+import { path } from 'ramda'
+import { useProduct } from 'vtex.product-context'
 import { Link, canUseDOM } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
 import ShowMore from 'react-show-more'
@@ -34,15 +29,6 @@ import ReviewForm from './ReviewForm'
 import AppSettings from '../graphql/appSettings.graphql'
 import ReviewsByProductId from '../graphql/reviewsByProductId.graphql'
 import AverageRatingByProductId from '../graphql/averageRatingByProductId.graphql'
-
-interface Product {
-  productId: string
-  productName: string
-}
-
-interface Props {
-  client: ApolloClient<NormalizedCacheObject>
-}
 
 interface Review {
   id: number
@@ -314,12 +300,12 @@ const CSS_HANDLES = [
   'reviewCommentUser',
 ] as const
 
-const Reviews: FunctionComponent<InjectedIntlProps & Props> = props => {
-  const { client, intl } = props
-
+function Reviews() {
+  const client = useApolloClient()
+  const intl = useIntl()
   const handles = useCssHandles(CSS_HANDLES)
-  const { product } = useContext(ProductContext) as any
-  const { productId, productName }: Product = product || {}
+  const { product } = useProduct() ?? {}
+  const { productId, productName } = product || {}
 
   const [state, dispatch] = useReducer(reducer, initialState)
 
@@ -742,4 +728,5 @@ const Reviews: FunctionComponent<InjectedIntlProps & Props> = props => {
   )
 }
 
-export default flowRight([withApollo, injectIntl])(Reviews)
+export default Reviews
+