+++ react/admin/components/util/productCellRenderer.tsx
@@ -5,8 +5,7 @@ import copy from 'copy-to-clipboard'
 import {
   FormattedMessage,
   defineMessages,
-  injectIntl,
-  InjectedIntlProps,
+  useIntl,
 } from 'react-intl'
 import { useQuery } from 'react-apollo'
 import {
@@ -23,7 +22,7 @@ const DEFAULT_TOAST_DURATION_MS = 1500
 
 type Props = {
   cellData: Product
-} & InjectedIntlProps
+}
 
 interface ProductResult {
   productName: string
@@ -49,7 +48,8 @@ const messages = defineMessages({
   },
 })
 
-const IntlProductCellRenderer: React.FC<Props> = ({ cellData, intl }) => {
+const ProductCellRenderer: React.FC<Props> = ({ cellData }) => {
+  const intl = useIntl()
   const [showCopy, setShowCopy] = useState(false)
 
   const { data } = useQuery<ProductData, ProductVars>(ProductQuery, {
@@ -130,7 +130,5 @@ const IntlProductCellRenderer: React.FC<Props> = ({ cellData, intl }) => {
     </ToastConsumer>
   ) : null
 }
-
-const ProductCellRenderer = injectIntl(IntlProductCellRenderer)
 
 export default ProductCellRenderer